+++ packages/s3-entities/src/getResolveJSONKey.ts
@@ -1,5 +1,0 @@
-import { Authority, Namespace, ObjectID } from "@phylopic/utils"
-import { encodeKeySegment } from "./encodeKeySegment"
-
-export const getResolveJSONKey = (build: number, authority: Authority, namespace: Namespace, objectID: ObjectID) =>
-    [build, "resolve", authority, namespace, objectID].map(encodeKeySegment).join("/") + ".json"